Added GPL Header to all .vala files
I believe that every file in the application needs to have the GPL
license. I have read the Elementary OS documentation and reviewed
the rules layed forth by the GNU people. In the Elementary OS docs,
there is no such instructions, but according to the GNU rules,
every file needs to have a copyright notice.