Keep Sidebar Visible Option (#85)
* Added keep sidebar visible switchbox

This feature is to keep the sidebar (Notebook area) from
disappearing when in EDIT mode.

* Fixed formating of if-statement

* Fixed if-statement format

* Hide sidebar when in toggle sidebar switchbox

When the user is in EDIT mode, if he or she toggles the switch
then the sidebar will appear or hide. This will make it apparent
to the user of whether the sidebar will remain visible.